Rename Content to Switcher.
